Final Project 3.0

+++ database_service/app/main.py
@@ -2,9 +2,10 @@ from fastapi import FastAPI, Depends
 from sqlalchemy.orm import Session
 from fastapi import HTTPException, status
 
-from database_service.database import engine, Base, SessionLocal
-from database_service.models import Cliente
-from database_service.schemas import ClienteCreate, ClienteUpdate
+from database_service.database.database import engine, Base, SessionLocal
+from database_service.models.models import Cliente
+from database_service.schemas.schemas import ClienteCreate, ClienteUpdate, ClienteUpdateUnit
+from cliente_service.service.service import calcular_score
 
 
 app = FastAPI()
@@ -98,6 +99,44 @@ def atualizar_cliente(cliente_id: int, cliente: ClienteUpdate, db: Session = Dep
 
 
 
+@app.patch("/clientes/{cliente_id}", status_code=status.HTTP_200_OK)
+def atualizar_cliente_parcial(cliente_id: int, cliente: ClienteUpdateUnit, db: Session = Depends(get_db)):
+    cliente_db = db.query(Cliente).filter(Cliente.id == cliente_id).first()
+
+    if not cliente_db:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Cliente não encontrado"
+        )
+
+    if cliente.nome is not None:
+        cliente_db.nome = cliente.nome
+
+    if cliente.telefone is not None:
+        cliente_db.telefone = cliente.telefone
+
+    if cliente.correntista is not None:
+        cliente_db.correntista = cliente.correntista
+
+    if cliente.score_credito is not None:
+        cliente_db.score_credito = cliente.score_credito
+
+    if cliente.saldo_cc is not None:
+        cliente_db.saldo_cc = cliente.saldo_cc
+
+    if cliente_db.correntista is False:
+        cliente_db.saldo_cc = 0.0
+
+    if cliente_db.saldo_cc <= 0:
+        cliente_db.score_credito = 0.0
+    else:
+        score_credito = calcular_score(cliente_db.saldo_cc)
+        cliente_db.score_credito = score_credito
+
+    db.commit()
+
+    return {"mensagem": "Cliente atualizado com sucesso"}
+
 
 
 @app.delete("/clientes/{cliente_id}", status_code=status.HTTP_200_OK)
@@ -109,6 +148,12 @@ def deletar_cliente(cliente_id: int, db: Session = Depends(get_db)):
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Cliente não encontrado"
+        )
+
+    if cliente.saldo_cc != 0:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Não é possível deletar um cliente com saldo na conta corrente"
         )
 
     db.delete(cliente)
